CourseDTO and validations

+++ src/main/java/app/services/CourseService.java
@@ -1,5 +1,6 @@
 package app.services;
 
+import java.time.ZonedDateTime;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,6 +8,7 @@ import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import app.domain.Course;
 import app.domain.User;
+import app.domain.dto.CourseDTO;
 import app.domain.enums.RoleType;
 import app.exceptions.ResourceNotFoundException;
 import app.repositories.CourseRepository;
@@ -29,30 +31,48 @@ public class CourseService {
 		return courseRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException());
 	}
 
-	public Course create(Course obj) {
-		//Verificar se quem está criando um professor ou admin
-		
+	public Course create(CourseDTO obj) {
+		//Apenas professores e admins podem criar cursos
 		User user = authenticatedUser.getAuthenticatedUser();
 		
 		if (user.getRole().equals(RoleType.STUDENT))
 			throw new AccessDeniedException("You have no permission to access this endpoint.");
 		
-		return courseRepository.save(obj);
+		Course course = new Course(null, obj.title(), obj.description(), user, ZonedDateTime.now().toLocalDate());
+		
+		return courseRepository.save(course);
 	}
 
-	public Course update(Course obj, Long id) {
+	public Course update(CourseDTO obj, Long id) {
 		//Verificar se quem está atualizando é o professor ou admin
+		User user = authenticatedUser.getAuthenticatedUser();
+		
+		if (user.getRole().equals(RoleType.STUDENT))
+			throw new AccessDeniedException("You have no permission to access this endpoint.");
+		
 		Course existingCourse = this.findById(id);
 		
-		existingCourse.setTitle(obj.getTitle());
-		existingCourse.setDescription(obj.getDescription());
-		existingCourse.setTeacher(obj.getTeacher());
+		if (user.getRole().equals(RoleType.TEACHER) && !user.getId().equals(existingCourse.getTeacher().getId()))
+			throw new AccessDeniedException("You can't update other teacher courses");
+		
+		existingCourse.setTitle(obj.title());
+		existingCourse.setDescription(obj.description());
 		
 		return courseRepository.save(existingCourse);
 	}
 
 	public void delete(Long id) {
 		//Verificar se quem está deletando é o professor ou admin
+		User user = authenticatedUser.getAuthenticatedUser();
+		
+		if (user.getRole().equals(RoleType.STUDENT))
+			throw new AccessDeniedException("You have no permission to access this endpoint.");
+		
+		Course existingCourse = this.findById(id);
+		
+		if (user.getRole().equals(RoleType.TEACHER) && !user.getId().equals(existingCourse.getTeacher().getId()))
+			throw new AccessDeniedException("You can't delete other teacher courses");
+		
 		courseRepository.deleteById(id);
 	}
